+++ src/main/java/insane96mcp/insanesurvivaloverhaul/setup/Client.java
@@ -114,8 +114,6 @@ public class Client {
         else if (event.getTabKey() == CreativeModeTabs.FOOD_AND_DRINKS) {
             if (Feature.isEnabled(FoodDrinks.class)) {
                 addAfter(event, Items.MUSHROOM_STEW, FoodDrinks.FUNGI_STEW);
-                addAfter(event, Items.MUSHROOM_STEW, FoodDrinks.BROWN_MUSHROOM_STEW);
-                addAfter(event, Items.MUSHROOM_STEW, FoodDrinks.RED_MUSHROOM_STEW);
                 addAfter(event, Items.COOKIE, FoodDrinks.OVER_EASY_EGG);
                 addBefore(event, Items.PUMPKIN_PIE, FoodDrinks.PUMPKIN_PULP);
                 addAfter(event, Items.PUMPKIN_PIE, FoodDrinks.APPLE_PIE);